requirements.txt
aioquic-mitmproxy==0.9.20.3
amqp==5.1.1
anyio==3.7.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-lru==2.0.2
asyncpg==0.28.0
attrs==23.1.0
autobahn==23.6.1
Automat==22.10.0
Babel==2.12.1
backcall==0.2.0
baidu-aip==4.16.12
beautifulsoup4==4.12.2
bidict==0.22.1
billiard==4.1.0
bleach==6.0.0
blinker==1.6.2
breadability==0.1.20
Brotli==1.0.9
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
comm==0.1.3
constantly==15.1.0
contourpy==1.0.7
cryptography==41.0.1
cssselect==1.2.0
cycler==0.11.0
daphne==4.0.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.2
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-crontab==0.7.1
django-import-export==3.2.0
django-json-widget==1.1.1
django-otp==1.5.4
django-simpleui==2023.11.16
django-tinymce==3.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
drf-spectacular==0.26.3
drf-spectacular-sidecar==2023.7.1
drf-yasg==1.21.8
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==1.2.0
fake-useragent==1.2.1
fastjsonschema==2.17.1
filelock==3.12.1
Flask==2.3.2
fonttools==4.39.4
formulas==1.2.7
fqdn==1.5.1
future==0.18.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incdbscan==0.1.0
incremental==22.10.0
inflection==0.5.1
ipykernel==6.24.0
ipython==8.14.0
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.14
jsonlines==4.0.0
jsonpointer==2.4
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.23.0
kaitaistruct==0.10
kiwisolver==1.4.4
kombu==5.3.1
ldap3==2.9.1
lxml==4.9.2
MarkupPy==1.14
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.1
mitmproxy==10.0.0
mitmproxy_rs==0.2.1
msgpack==1.0.5
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==2.8.8
nltk==3.8.1
notebook_shim==0.2.3
numpy==1.24.3
numpy-financial==1.0.0
odfpy==1.4.1
opencv-python==4.7.0.72
openpyxl==3.1.2
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parsel==1.8.1
parso==0.8.3
passlib==1.7.4
pdfminer.six==20221105
pdfplumber==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
prometheus-client==0.17.0
prompt-toolkit==3.0.39
Protego==0.2.1
protobuf==4.23.4
psutil==5.9.5
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg2-binary==2.9.6
ptyprocess==0.7.0
publicsuffix2==2.20191221
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==22.3.5
pycparser==2.21
pycryptodome==3.18.0
PyDispatcher==2.0.7
PyExecJS==1.5.1
Pygments==2.15.1
PyJWT==2.7.0
pylsqpack==0.3.17
PyMuPDF==1.22.3
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.19.3
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
queuelib==1.6.2
regex==2023.6.3
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
schedula==1.5.2
scikit-learn==1.5.1
scipy==1.14.0
Scrapy==2.9.0
scrapy-djangoitem==1.1.1
scrapyd==1.4.2
scrapyd-client==1.2.3
Send2Trash==1.8.2
sentry-sdk==1.25.1
service-identity==21.1.0
six==1.16.0
sklearn==0.0.post5
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
sumy==0.11.0
tablib==3.5.0
terminado==0.17.1
threadpoolctl==3.5.0
tinycss2==1.2.1
tldextract==3.4.4
tomli==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
uberegg==0.1.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.3
urwid-mitmproxy==2.1.2.1
vine==5.0.0
w3lib==2.1.1
Wand==0.6.11
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
xxhash==2.0.2
zope.interface==6.0
zstandard==0.21.0
